Ensure an errors.ExitCoder is returned in run
When the value returned from Validate or Execute was just an error nothing would
be printed to the console. Check to see if the value returned by the plugin is an
ExitCoder and if not create an ExitCoder to ensure the user is notified as to what
happened during the run.